# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reservations - API",
        "version": "1.0"
    },
    "paths": {
        "/api/cinemas": {
            "get": {
                "tags": [
                    "Cinemas"
                ],
                "summary": "Gets the list of cinemas",
                "description": "Returns the list of cinemas with optional pagination",
                "operationId": "getCinemasList",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of cinemas per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No cinema(s) found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cinemas"
                ],
                "summary": "Store a new cinema",
                "description": "Stores a new cinema and returns the cinema data",
                "operationId": "storeCinema",
                "requestBody": {
                    "description": "Cinema data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "A New Beginning"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "A New Beginning"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/cinemas/{id}": {
            "get": {
                "tags": [
                    "Cinemas"
                ],
                "summary": "Get cinema by ID",
                "description": "Returns a single cinema",
                "operationId": "getCinemaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cinema to return",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "A New Beginning"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cinema not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cinemas"
                ],
                "summary": "Update an existing cinema",
                "description": "Updates and returns a cinema data",
                "operationId": "updateCinema",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cinema that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Cinema data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "A New Beginning"
                                    },
                                    "created_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-09-15"
                                    },
                                    "updated_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2021-09-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cinema updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "3434-4343-4343-4343"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "A New Beginning"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-09-15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Cinema not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cinemas"
                ],
                "summary": "Delete a cinema",
                "description": "Deletes a cinema and returns a success message",
                "operationId": "deleteCinema",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cinema to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Cinema deleted successfully"
                    },
                    "404": {
                        "description": "Cinema not found"
                    },
                    "500": {
                        "description": "Error deleting movie"
                    }
                }
            }
        },
...
    }
}